requirements.txt
Django>=3.1,<3.2
wagtail>=2.11,<2.12
django-debug-toolbar==3.2
django-allauth==0.44.0
psycopg2-binary==2.8.6
django-extensions==3.1.0